requirements.txt
flask==2.3.3
flask-mysqldb==1.0.1
python-dotenv==1.0.0